refactor steps to separate files
